Rename Method Changes, Utility Added -Added a success confirmation to the rename files method telling the user that the rename was successful. I also added a counter and show how many files were renamed. -I used the currentIdentifier (current 5-letter TDS identifier) entered by the user to loop through the files before attempting to rename and ensure that files matchiing that pattern do exist. If not, the method exits. -Moved the Environment.GetLogicalDrives method into a helper method that first cleared the comboBox that stored the dirves, and repopulated it. I then call this helper method from the click even of the comboBox so that everytime the box is clicked, the drives update. I did this in case a user inserts a flash/external drive after the program is running. Previously, the Environment.GetLogicalDrives method was only call in the Form_Load event.
Utilities
-In the utilities class, I added a simple message box that will pop up when the user tries to close out the program, asking for exit confirmation. I put it in the utilities class for reuseability. Just call Utilities.closeConfirmation in the Form_Closing event for the current form, and pass it the sender and e variables. This will allow for close confirmation across forms.

+++ TDSDataReader2/TestBuild/fileRenameForm.cs
@@ -22,8 +22,27 @@ namespace TDSDataReader2.TestBuild
 
         private void fileRenameForm_Load(object sender, EventArgs e)
         {
-            //Populate the drivesComboBox with valid drives
+            CenterToScreen();
+            //initially populate the drives into the comboBox for selection
+            populateDriveCombo();            
+        }
+
+        //Call a close conformation box to confirm user wants to close the program.
+        private void fileRenameForm_FormClosing(object sender, FormClosingEventArgs e)
+        {
+            Utilities.closeConfirmation(sender, e);
+        }
+
+        //Populate the drivesComboBox with valid drives. Resets in case drive has been added.
+        protected void populateDriveCombo()
+        {
+            drivesComboBox.Items.Clear();
             drivesComboBox.Items.AddRange(Environment.GetLogicalDrives());
+        }
+
+        private void drivesComboBox_Click(object sender, EventArgs e)
+        {
+            populateDriveCombo();
         }
 
         //Gets the file path to use for renaming the files
@@ -70,25 +89,43 @@ namespace TDSDataReader2.TestBuild
             }
         }
 
-        //Rename the files according to the input from the user
-        //Still need to add some features to this, such as try/catch, confirmation dialogue box (to give user the chance to cancel) and a better success message box.
+        //Rename the files according to the input from the user. Confirmation dialogue pops up asking user if they are sure of changes. If yes, file rename runs.
         protected void renameFiles(string path)
         {
-            DirectoryInfo dir = new DirectoryInfo(path);
-            FileInfo[] files = dir.GetFiles();
-            foreach (FileInfo file in files)
+            if (MessageBox.Show("Rename " + currentIdentifier + "*.* " + "to " + newIdentifier + "*.*?" + "\n\n" + "ARE YOU SURE THESE CHANGES ARE CORRECT?", "Confirm", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
             {
-                File.Move(file.FullName, file.FullName.ToString().Replace(currentIdentifier, newIdentifier));
+                int counter = 0;
+
+                DirectoryInfo dir = new DirectoryInfo(path);
+                //Make sure that the files exist. If so, run the file rename and show success message. If not, show error message.
+                var checkFiles = dir.GetFiles(currentIdentifier + "*.*");
+                if (checkFiles.Length > 0) // files exist
+                {
+                    //Place files into array and loop through renaming them
+                    FileInfo[] files = dir.GetFiles();
+                    foreach (FileInfo file in files)
+                    {
+                        File.Move(file.FullName, file.FullName.ToString().Replace(currentIdentifier, newIdentifier));
+                        counter++;
+                    }
+
+                    MessageBox.Show("Successfully renamed " + counter + " files from " + currentIdentifier + "*.* " + "to " + newIdentifier + "*.*", "Success!");
+
+                    clearData();
+                    chooseDriveRadio.Focus();
+                }
+                else //files do not exist
+                {
+                    MessageBox.Show("There are no files in this directory that match your criteria. Please try again.", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    clearData();
+                }
             }
-            MessageBox.Show("Success!");
-            clearData();
         }
 
         //Clear all data from the controls on the form
         protected void clearData()
         {
             Utilities.ClearControls(this);
-
         }
 
         private void chooseDriveRadio_CheckedChanged(object sender, EventArgs e)
@@ -160,6 +197,5 @@ namespace TDSDataReader2.TestBuild
                 else renameFiles(pathToTdsFiles);
             }
         }
-
     }
 }
